Move zip-file wrapper to release-included module
